// src/app/data/generatedContentMedia.ts
export const generatedConstructorFlowers: ConstructorFlowerAsset[] = [
  {
    "id": "constructor-rose-001",
    "name": "Аваланж Белая",
    "image": "/products/constructor/avalanzh-belaya/1.webp",
    "images": [
      "/products/constructor/avalanzh-belaya/1.webp"
    ]
  },
// ...
];